Split app into py2 and py3 services.
