+++ internals/notifier_test.py
@@ -23,7 +23,6 @@ import flask
 import mock
 import werkzeug.exceptions  # Flask HTTP stuff.
 
-from google.appengine.api import mail
 # TODO(jrobbins): phase out gae_users.
 from google.appengine.api import users as gae_users
 from framework import users
@@ -335,154 +334,3 @@ class FeatureStarTest(testing_config.CustomTestCase):
     self.assertItemsEqual(
         [app_user_1.email, app_user_2.email],
         [au.email for au in actual])
-
-
-class OutboundEmailHandlerTest(testing_config.CustomTestCase):
-
-  def setUp(self):
-    self.handler = notifier.OutboundEmailHandler()
-    self.request_path = '/tasks/outbound-email'
-
-    self.to = 'user@example.com'
-    self.subject = 'test subject'
-    self.html = '<b>body</b>'
-    self.sender = ('Chromestatus <admin@%s.appspotmail.com>' %
-                   settings.APP_ID)
-
-  @mock.patch('settings.SEND_EMAIL', True)
-  @mock.patch('settings.SEND_ALL_EMAIL_TO', None)
-  @mock.patch('google.appengine.api.mail.EmailMessage')
-  def test_post__prod(self, mock_emailmessage_constructor):
-    """On cr-status, we send emails to real users."""
-    params = {
-        'to': self.to,
-        'subject': self.subject,
-        'html': self.html,
-        }
-    with notifier.app.test_request_context(self.request_path, json=params):
-      actual_response = self.handler.process_post_data()
-
-    mock_emailmessage_constructor.assert_called_once_with(
-        sender=self.sender, to=self.to, subject=self.subject,
-        html=self.html)
-    mock_message = mock_emailmessage_constructor.return_value
-    mock_message.check_initialized.assert_called_once_with()
-    mock_message.send.assert_called_once_with()
-    self.assertEqual({'message': 'Done'}, actual_response)
-
-  @mock.patch('settings.SEND_EMAIL', True)
-  @mock.patch('google.appengine.api.mail.EmailMessage')
-  def test_post__staging(self, mock_emailmessage_constructor):
-    """On cr-status-staging, we send emails to an archive."""
-    params = {
-        'to': self.to,
-        'subject': self.subject,
-        'html': self.html,
-        }
-    with notifier.app.test_request_context(self.request_path, json=params):
-      actual_response = self.handler.process_post_data()
-
-    expected_to = 'cr-status-staging-emails+user+example.com@google.com'
-    mock_emailmessage_constructor.assert_called_once_with(
-        sender=self.sender, to=expected_to, subject=self.subject,
-        html=self.html)
-    mock_message = mock_emailmessage_constructor.return_value
-    mock_message.check_initialized.assert_called_once_with()
-    mock_message.send.assert_called_once_with()
-    self.assertEqual({'message': 'Done'}, actual_response)
-
-  @mock.patch('settings.SEND_EMAIL', False)
-  @mock.patch('google.appengine.api.mail.EmailMessage')
-  def test_post__local(self, mock_emailmessage_constructor):
-    """When running locally, we don't actually send emails."""
-    params = {
-        'to': self.to,
-        'subject': self.subject,
-        'html': self.html,
-        }
-    with notifier.app.test_request_context(self.request_path, json=params):
-      actual_response = self.handler.process_post_data()
-
-    expected_to = 'cr-status-staging-emails+user+example.com@google.com'
-    mock_emailmessage_constructor.assert_called_once_with(
-        sender=self.sender, to=expected_to, subject=self.subject,
-        html=self.html)
-    mock_message = mock_emailmessage_constructor.return_value
-    mock_message.check_initialized.assert_called_once_with()
-    mock_message.send.assert_not_called()
-    self.assertEqual({'message': 'Done'}, actual_response)
-
-
-class BouncedEmailHandlerTest(testing_config.CustomTestCase):
-
-  def setUp(self):
-    self.handler = notifier.BouncedEmailHandler()
-    self.sender = ('Chromestatus <admin@%s.appspotmail.com>' %
-                   settings.APP_ID)
-    self.expected_to = settings.BOUNCE_ESCALATION_ADDR
-
-  @mock.patch('internals.notifier.BouncedEmailHandler.receive')
-  def test_process_post_data(self, mock_receive):
-    with notifier.app.test_request_context('/_ah/bounce'):
-      actual_json = self.handler.process_post_data()
-
-    self.assertEqual({'message': 'Done'}, actual_json)
-    mock_receive.assert_called_once()
-
-
-  @mock.patch('settings.SEND_EMAIL', True)
-  @mock.patch('google.appengine.api.mail.EmailMessage')
-  def test_receive__user_has_prefs(self, mock_emailmessage_constructor):
-    """When we get a bounce, we update the UserPrefs for that user."""
-    starrer_3_pref = models.UserPref(
-        email='starrer_3@example.com',
-        notify_as_starrer=False)
-    starrer_3_pref.put()
-
-    bounce_message = testing_config.Blank(
-        original={'to': 'starrer_3@example.com',
-                  'from': 'sender',
-                  'subject': 'subject',
-                  'text': 'body'})
-
-    self.handler.receive(bounce_message)
-
-    updated_pref = models.UserPref.get_by_id(starrer_3_pref.key.integer_id())
-    self.assertEqual('starrer_3@example.com', updated_pref.email)
-    self.assertTrue(updated_pref.bounced)
-    self.assertFalse(updated_pref.notify_as_starrer)
-    expected_subject = "Mail to 'starrer_3@example.com' bounced"
-    mock_emailmessage_constructor.assert_called_once_with(
-        sender=self.sender, to=self.expected_to, subject=expected_subject,
-        body=mock.ANY)
-    mock_message = mock_emailmessage_constructor.return_value
-    mock_message.check_initialized.assert_called_once_with()
-    mock_message.send.assert_called()
-
-  @mock.patch('settings.SEND_EMAIL', True)
-  @mock.patch('google.appengine.api.mail.EmailMessage')
-  def test_receive__create_prefs(self, mock_emailmessage_constructor):
-    """When we get a bounce, we create the UserPrefs for that user."""
-    # Note, no existing UserPref for starrer_4.
-
-    bounce_message = testing_config.Blank(
-        original={'to': 'starrer_4@example.com',
-                  'from': 'sender',
-                  'subject': 'subject',
-                  'text': 'body'})
-
-    self.handler.receive(bounce_message)
-
-    prefs_list = models.UserPref.get_prefs_for_emails(
-        ['starrer_4@example.com'])
-    updated_pref = prefs_list[0]
-    self.assertEqual('starrer_4@example.com', updated_pref.email)
-    self.assertTrue(updated_pref.bounced)
-    self.assertTrue(updated_pref.notify_as_starrer)
-    expected_subject = "Mail to 'starrer_4@example.com' bounced"
-    mock_emailmessage_constructor.assert_called_once_with(
-        sender=self.sender, to=self.expected_to, subject=expected_subject,
-        body=mock.ANY)
-    mock_message = mock_emailmessage_constructor.return_value
-    mock_message.check_initialized.assert_called_once_with()
-    mock_message.send.assert_called()
