+++ scripts/deploy_site.sh
@@ -29,5 +29,7 @@ gcloud app deploy \
   --project $appName \
   --version $deployVersion \
   --no-promote \
-  $BASEDIR/../app.yaml $BASEDIR/../notifier.yaml \
+  $BASEDIR/../app-py2.yaml $BASEDIR/../notifier.yaml \
   $BASEDIR/../dispatch.yaml $BASEDIR/../cron.yaml
+
+# TODO(jrobbins): Add app_py3.yaml when it has some contents.
