+++ scripts/start_server.sh
@@ -11,4 +11,11 @@ readonly BASEDIR=$(dirname $BASH_SOURCE)
 dev_appserver.py -A cr-status --enable_console=1 \
   --support_datastore_emulator=1 --datastore_emulator_port=15606 \
   --env_var DATASTORE_EMULATOR_HOST='localhost:15606' \
-  $BASEDIR/../app.yaml $BASEDIR/../notifier.yaml
+  $BASEDIR/../dispatch.yaml \
+  $BASEDIR/../app-py2.yaml $BASEDIR/../notifier.yaml \
+  $BASEDIR/../app-py3.yaml
+
+# TODO(jrobbins): Add app-py3.yaml when it has some contents.  At some point in
+# the future when we cannot run the app-py2 service locally, tasks for
+# outbound-email would simply queue up forever.  However, we don't even create
+# those tasks in dev mode.
